requirements/base.txt
pytz==2021.1  # https://github.com/stub42/pytz
python-slugify==5.0.2  # https://github.com/un33k/python-slugify
Pillow==8.2.0  # https://github.com/python-pillow/Pillow
argon2-cffi==20.1.0  # https://github.com/hynek/argon2_cffi
whitenoise==5.2.0  # https://github.com/evansd/whitenoise
redis==3.5.3  # https://github.com/andymccurdy/redis-py
hiredis==1.1.0  # https://github.com/redis/hiredis-py

# Django
# ------------------------------------------------------------------------------
django==3.1.11  # pyup: < 3.2  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==4.1.1  # https://github.com/jazzband/django-model-utils
django-allauth==0.44.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.11.2  # https://github.com/django-crispy-forms/django-crispy-forms
django-redis==4.12.1  # https://github.com/jazzband/django-redis
six==1.15.0
reportlab==3.6.1
#Pillow==8.2.0
#numpy==1.19.4
imagesize==1.2.0
django-widget-tweaks==1.4.8
# pandas==1.1.5
django-formtools==2.3
django-filter==2.4.0
django-xhtml2pdf==0.0.4
pyPdf2==1.26.0
arabic-reshaper==2.1.3
future==0.18.2
html5lib==1.1
python-bidi==0.4.2
webencodings==0.5.1
xhtml2pdf==0.2.5
django-pwa==1.0.10
